requirements.txt
Django==4.1.3
djangorestframework==3.14.0
psycopg2-binary==2.9.4
flake8>=3.9.2,<3.10
Pillow==9.4.0
